# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "同桌小程序",
        "title": "Deskmat",
        "termsOfService": "http://swagger.io/terrms/",
        "contact": {
            "name": "qianren",
            "email": "1911401642@qq.com"
        },
        "version": "1.0.0"
    },
    "host": "119.3.2.168:4016",
    "basePath": "/api/v1/",
    "paths": {
        "/apply": {
            "get": {
                "description": "点击 同桌申请 时调用",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apply"
                ],
                "summary": "显示收到的同桌申请",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"success\"}"
                    },
                    "400": {
                        "description": "{\"error_code\":\"00001\", \"message\":\"Fail.\"} or {\"error_code\":\"00002\", \"message\":\"Lack Param Or Param Not Satisfiable.\"}"
                    },
                    "401": {
                        "description": "{\"error_code\":\"10001\", \"message\":\"Token Invalid.\"} 身份认证失败 重新登录"
                    },
                    "500": {
                        "description": "{\"error_code\":\"30001\", \"message\":\"Fail.\"} 失败"
                    }
                }
            },
            "put": {
                "description": "用户确认接受的同桌申请",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apply"
                ],
                "summary": "确认同桌申请,进入同桌打卡",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "要同意申请的同学学号",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Respondent"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"success\"}"
                    },
                    "401": {
                        "description": "{\"msg\":\"confirm faided\"}"
                    }
                }
            },
            "post": {
                "description": "从名片页面像对方发出申请",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apply"
                ],
                "summary": "同桌申请",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "申请对象的学号(id)",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Respondent"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"success\", \"对象id\":\"string\"}"
                    },
                    "400": {
                        "description": "{\"error_code\":\"00001\", \"message\":\"Fail.\"} or {\"error_code\":\"00002\", \"message\":\"Lack Param Or Param Not Satisfiable.\"}"
                    },
                    "401": {
                        "description": "{\"error_code\":\"10001\", \"message\":\"Token Invalid.\"} 身份认证失败 重新登录"
                    },
                    "500": {
                        "description": "{\"error_code\":\"30001\", \"message\":\"Fail.\"} 失败"
                    }
                }
            }
        },
        "/apply/refuse": {
            "put": {
                "description": "用户拒绝接受的同桌申请",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apply"
                ],
                "summary": "拒绝同桌申请",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "要拒绝申请的同学学号",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Respondent"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"success\"}"
                    },
                    "401": {
                        "description": "{\"msg\":\"confirm faided\"}"
                    }
                }
            }
        },
        "/card": {
            "get": {
                "description": "\"获取自己的名片信息\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "card"
                ],
                "summary": "我的名片",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Card"
                        }
                    },
                    "401": {
                        "description": "身份验证失败"
                    },
                    "404": {
                        "description": "获取失败"
                    }
                }
            },
            "put": {
                "description": "\"修改名片的昵称，标签，头像，用户宣言\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "card"
                ],
                "summary": "修改名片信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "名片信息",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Cardinfor"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "修改成功"
                    },
                    "400": {
                        "description": "修改失败"
                    },
                    "401": {
                        "description": "验证失败"
                    }
                }
            },
            "post": {
                "description": "\"设置名片的昵称，标签，头像，用户宣言\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "card"
                ],
                "summary": "设置名片",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "名片信息",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Cardinfor"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "设置成功"
                    },
                    "400": {
                        "description": "设置失败"
                    },
                    "401": {
                        "description": "身份验证失败 重新登录"
                    }
                }
            }
        },
...
        "/card/infor": {
            "post": {
                "description": "\"获取他人的名片信息\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "card"
                ],
                "summary": "他人名片",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "要查看用户的学号",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Card"
                        }
                    },
                    "401": {
                        "description": "身份验证失败"
                    },
                    "404": {
                        "description": "获取失败"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.Card": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "college": {
                    "description": "学院",
                    "type": "string"
                },
                "declaration": {
                    "type": "string"
                },
                "grade": {
                    "description": "年级",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "infor": {
                    "type": "string"
                },
                "major": {
                    "description": "专业",
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "sha": {
                    "type": "string"
                },
                "status": {
                    "description": "记录是否有同桌",
                    "type": "string"
                },
                "tag1": {
                    "type": "string"
                },
                "tag2": {
                    "type": "string"
                },
                "tag3": {
                    "type": "string"
                },
                "tag4": {
                    "type": "string"
                },
                "tag5": {
                    "type": "string"
                },
                "users_id": {
                    "type": "string"
                }
            }
        },
        "model.Cardinfor": {
            "type": "object",
            "properties": {
                "college": {
                    "type": "string"
                },
                "declaration": {
                    "type": "string"
                },
                "grade": {
                    "type": "string"
                },
                "infor": {
                    "type": "string"
                },
                "major": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "tag1": {
                    "type": "string"
                },
                "tag2": {
                    "type": "string"
                },
                "tag3": {
                    "type": "string"
                },
                "tag4": {
                    "type": "string"
                },
                "tag5": {
                    "type": "string"
                }
            }
        },
        "model.Id": {
            "type": "object",
            "properties": {
                "users_id": {
                    "type": "string"
                }
            }
        },
...
        "model.Respondent": {
            "type": "object",
            "properties": {
                "respondent_id": {
                    "type": "string"
                }
            }
        },
...
    }
}